extracted handler from index
