+++ integrations/sunco/src/index.ts
@@ -1,8 +1,7 @@
 import { sentry as sentryHelpers } from '@botpress/sdk-addons'
 import { actions } from './actions'
 import { channels } from './channels'
-import { executeConversationCreated, handleConversationMessage } from './events'
-import { isSuncoWebhookPayload } from './messaging-events'
+import { handler } from './handler'
 import { register, unregister } from './setup'
 import * as bp from '.botpress'
 
@@ -11,29 +10,7 @@ const integration = new bp.Integration({
   unregister,
   actions,
   channels,
-  handler: async ({ req, client, logger }) => {
-    if (!req.body) {
-      console.warn('Handler received an empty body')
-      return
-    }
-
-    const data = JSON.parse(req.body)
-
-    if (!isSuncoWebhookPayload(data)) {
-      logger.forBot().warn('Received an invalid payload from Sunco')
-      return
-    }
-
-    for (const event of data.events) {
-      if (event.type === 'conversation:create') {
-        await executeConversationCreated({ event, client, logger })
-      } else if (event.type === 'conversation:message') {
-        await handleConversationMessage(event, client, logger)
-      } else {
-        console.warn(`Received an event of type ${event.type}, which is not supported`)
-      }
-    }
-  },
+  handler,
 })
 
 export default sentryHelpers.wrapIntegration(integration, {
